do not automatically build the uri_name on the resource save
Make the uri_name association setup explicit as the User and UserGroup models are inherently coupled to the uri_name model. The coupling is not ideal but there is no other way to ensure uniq names between users and groups without the uri_name model in the system. Care should be taken when setting up the resource models.