add persistent data container for panoptes
This needs to be improved to get the dev and test gems installed to the underlying container on build. However since one can’t ref the Gemfile in the scripts/fig/Dockerfile via ADD ../../Gemfile then call bundle install we need another approach. 

I think switching the root docker file before first build to include the dev and test gems would suffice but this would require a rebuild of the base image when installing new gems. To be continued…