Add dependent relationships to models
When a project is destroyed its subject sets and associated models
should also be destroyed, but we should keep subjects around in case
they are being used by other projects. Classifications and workflows
should also remain so we can still redo the since if necessary.

Subjects and Workflows can still be explicitly destroyed on their own.

Closes #652