Dequeue and update user seen subjects in sync
Handle this bit of bookkeeping during the request/response cycle to make
sure seen subjects are never shown to users. Also apply the "uniq"
function in PostgreSQL to queues and seen lists to make sure they don't
contain duplicate data.

Enable the "intarray" extension in PostgreSQL, which is installed in
RDS, in order to make sure these array operations and faster and easier to understand.

This should Close zooniverse/Panoptes-Front-End#426 and zooniverse/Panoptes-Front-End#425