extract dequeue to high priority background worker
Avoid the case where lots of concurrent non-logged in users try to run multiple dequeues. Dequeues can happen in any order so we can background this andnot block the subject selector. The only problem with this is when these jobs queue up in sidekiq and don't remove from the queue for active users. The queue context should provide some UI feedback to the clients that should alleviate this.