Set finished_at when workflow finishes
Instead of deactivating, set a timestamp column. This allows us to see
which workflows are supposed to be visible but have just run out of
data. In turn, that will allow us to automatically unfinish a workflow
when more data becomes available (either due to recounts or simply
because more data is added to a subject set).