conditionally eager load project owner
the owner association is restrictive for joins and can not be eager loaded when filtering on current_user_roles that contain non-owner roles. I'm not a huge fan of this implementation but i couldn't get the join to be less restrictive whilst ensuring only the correct ACLs are applied for the param scope.  

An alternative would be to just drop the eager load for this owner association.