Avoid multiple classifcation counts on classification receipt (#1946)
* add a rate limited sws classifications_count worker

this will be used to update the counts to their correct value in the databse and rate limited to try and avoid contention on the locking updates to the counter column

* use a higher queue than the retirement worker

try and resolve the correct count before retirement runs

* use incremement_counter and a rate limited worker for classification counts

avoid the cost of recounting on the classifications table each time on classification receipt. Use the incrememnt the counter as an estimate (for reporting counts) and schedule a higher priority rate limited worker to update the correct count.