Add simplified completeness calculation for never_retire
Projects that use never_retire will usually still have some kind of
retirement, but it will be externally managed (Nero, or custom).

In this case, since we can't tell how many classifications we will need,
we can only use the number of retired subjects as a guide to how far
along a project is.