Subject remove default scope and optimize the selector scope queries (#2061)
* remove the subjects locations eager load

* add indexes for resource scope_for queries

add default private/public scope and activated_state indexes for controlled resources

* subjects preloading and flag to disable selection lookups

preload the n+1 subejct associations and provide a mechanism to turn off the possibly costly lookups for selected context, e.g. user finished workflow, already seen, is retired for the workflow when we run peaky traffic events.

* use a custom context object to switch the serializer behaviour

inject the dependency to the serializer instead of always paying the lookup cost of flipper in the serializer.

* simplify the custom subject selector scope

do not use the default controlled scope in the serializer for subject selection as it uses the parental controlled project scope and adds complexity to the final resolved scope. The custom selector will deterine the specific scope context (subject_ids, active, etc) so passing a basic scope will simplify the serliizer lookup / count queries

* remove Subject.unscoped contexts

as we no longer have the default eager load scope

* avoid user_seen lookup if no user

* describe the correct serializer

* remove redundant else clause

thanks hound :)

* use default selection scope

but allow the scope to be injected as a dependency

* put user_seens into selection context

pay cost once per page of data and avoid n+1 queries on the user_seens

* use the default selection scope from controller

* avoid db lookups for linked location resources

use sort on preloaded subject locations and db order by lookups for non-loaded location resources.

* avoid extra db call to order the preloaded location resources