refactor the project#update action
moved the primary_contents and tag updates from the update_params to the update action to ensure these side effects rollback on failure via the transaction in the update yield block. 

Also ensure when updating these attributes to touch the project’s updated_at timestamp to invalidate cache keys / etags for this resource (used in conditional gets).

I’ve left todo notes in there as hooks to extract to service operations to update the project’s primary_content and tags instead of inlining in the projects controller.