compare weak etags for update / destroy actions precondition checks (#2224)
* compare weak etags for update / destroy actions precondition checks

Allow weak etag comparison if we receive a weak_etag server side for modifications, etc. This is due to adding gzip compression to nginx as it modifes the etag to be weak as this is correct etag spec. Rails is off spec with strong etags (changing in ver 5 to be weak) and this change allows the etags to not be rejected api side.

* extract etag validation to common controller method

allow overriden precondition checks to use a common comparison method.

* use different reg expression syntax to avoid escaping