provide correct paging urls with has_many relations (#2239)
* wire up the has many filtering params via custom restpack options class

don't allow the filters to create scope filters as we generate that scope via joins but we do want to activate the normal paging filter params. this new class will just construct the filter params for use in the url params

* wire up the has many filtering params via custom restpack options class

don't allow the filters to create scope filters as we generate that scope via joins but we do want to activate the normal paging filter params. this new class will just construct the filter params for use in the url params

* use an initializer instead of overloading the context param

* add class method to construct the paging scope

* extract spec to shared helper

* build the paging scope for use in the has many filter options paging override

* add comments about why this is custom code

explain why this code doesn't use super and has it's own options class

* setup has_many_filtering using customized serializer code and preload data

* add has_many filtering spec  for workflow

* reset the instance vars to avoid polluting other tests

* move collection serializer to panoptes restpack for has_many filtering