Faster classification saves (#2275)
* use IN query vs joins for private ACL in default scope for

this was quicker in my basic benchmarking:
x.report("TEST") do            
  (1..10000).each do              
    public_query(private_query(action, user, roles), public_flag)                
  end
end

      user     system      total        real
IN  6.920000   0.070000   6.990000 (  7.085182)

       user     system      total        real
JOIN  7.990000   0.070000   8.060000 (  8.145074)

* clarify the method intention

this method doesn’t find the public query but really combines the private and the public scopes to retrieve all resources a user can access.

* avoid issuing a count query on find resource lookups

when the relation is a single resource we don’t have to issue a count query to decide if it exists. this refactor splits up relation or error tests depending if the find arg is an array vs single param.

* assign FK id for belongs_to link relations

avoid double lookups when finding the relation to link, find the relation and where we can assign the FK. Default back to rails relation assignment methods.

* revert to just straight association assignment

dome models require validations on the associations if we set the FK without the loaded association resource then it looks this up again and we double loading costs. Easiest way to avoid ambiguity about the relation presence is to assign the found link relation directly.

* remove old belongs_to_many link

* do not validate the subjects on classification save

* use the refactored relations_or_error method