remove the was_update false default param
the was_update boolean was used to signal a no count on the subject workflow status counter. The lifecycle method should never call count worker until the classification is complete. So instead of checking the action here just allow the lifecycle to call this worker when it should increment a count.