add experiment to test better classification counts
test the impact of the additional non-index filters on the SubjectWorkflow classification counters, too much additional work on the db will be a no-no for these counters as they run all the time over a large table. Best to get some metrics for this change and decide how to go from there.