Better subject classification counts used for reporting & retirement. (#2636)
* only count complete classifications for retirement.

* simplify the query for complete lookup

* do not count gold standard classification for retirement

these should be used for comparative analysis to determine quality of annotations to known experts, we still want to hit the number of users for each subject ignoring any expert effort on a subject.

* ignore seen before classifications in counters

do not double count seen before classifications in counters, especially for retirement.

* add experiment to test better classification counts

test the impact of the additional non-index filters on the SubjectWorkflow classification counters, too much additional work on the db will be a no-no for these counters as they run all the time over a large table. Best to get some metrics for this change and decide how to go from there.