add project metadata config option to restrict self service data access (#2612)
* add project metadata config option to disable publishing to stream 

some projects may contain sensitive data that we need to control access to. this is a first step to avoid leaking that data to the stream and associated services. If this becomes widely needed it should be a column on the project with an index.

* change project config field to private_data

make the configuration name more relevant for other parts of the app.

* add model method to test for disabled data exports

* allow workflow data exports to be disabled

respect project setting to ensure data isn't published without due process

* add shared controller helper spec

test the api repsonds with forbidden http code and relevant error message

* only before not prepended controller actions

ensure these come after the logged in checks

* remove focus scope

* remove duplicate authorized_user let block

* clarify intent with renamed disabled exports opt & method

ensure the intent of keeping this data internal only to the panoptes db is signalled by the project config option and the method name.

* forbid panoptes only classification api exports