Updatable resources should yield before they save
Race conditions can occur if the yield is to a sidekiq worker or something that pulls from the database before the following save! can commit its transaction. In lieu of a callback, this will at least wait until the save! call finishes (successfully or not) before the yield is attempted.