Do not split when param is already an array (#2735)
This would happen if using ?current_user_role[]=a style params, which Rails
will automatically parse and turn into an array of strings. Rails adds an
Array#split method, which caused the .try to execute and return a nested array,
which in turn happened to work with PostgresExt, but not with the new syntax.