Automate pausing / unpausing projects when finishing / linking new data (#2646)
* add note about what the gold standard annotations table is

noting that this data is for specific projects and was sliced out of classification table and long term can go, not till the projects finish / port to the new feedback subject metadata mechanism though.

* spec the completeness transition to paused

* move a project from active to paused when it’s complete

* add notes about race conditions with count workers

* Revert "add note about what the gold standard annotations table is"

This reverts commit f59c2803689697b190823aa12a0e7993fc73bc93.

* add ordered workflow refresh state worker

worker to group other workers and run them in band to ensure we don't have state race conditions between workers and/or db transactions

* use the refresh workflow status worker after linkage

avoid race conditions between workers, serialise their execution to allow for consistent state as they run.

* call refresh workflow status worker after retire operation

* remove comment about update actions

this will be addressed in #2647

* inline the chain of processing

* use retry semantics vs dropping refresh status workers

lock and drop while the job is in the queue and rely on rescheduling to ensure we have up to date data. This is mostly to handle things like unselecting subject sets from workflows one by one with some delay.

* remove the comment about job locking

this job is only run through sidekiq middleware via the scheduled tasks. Ideally it’ll be in band only but in the meantime this can still run on the hour

* rejig the ordered chained worker specs

* move to string comparison instead of regex