move operation to use the workflow_id
avoid serialising the workflow resource from the db just to use the id that we have in params. The controlled_resources.exists? check still runs to ensure we are allowed to retire for this workflow, but that doesn’t instantiate any resources from the db so we can trust the authorization code and simply use the params in the operation.