fix paging when filtering on login and email for users
allow the filters to cascade into the serialiser to construct the meta href for paging. Care has to be taken here to ensure we don’t allow non-admins to activate the filter in the serialiser and that we also don’t pollute the custom AR scopes with the non-lowered params, hence the custom Restpack::Options class.