Ensure UserSeenSubjects are unique
On production we found that some users have more than one USS record for
a given workflow. This breaks Designator in particular, which will fail
to load its user seens cache for these users (in that workflow only).

To deploy make sure to run the rake task for cleanup first, otherwise
the migration will fail. There are only a couple of duplicates in the
database, so I guess that the chance of a new one appearing between the
rake task and running migrations is low enough that we don't need
anything more complicated than this.