Begin work to add last project to dormant mailer
When sending emails to dormant users it would be good to include a link
to the last project they've classified on (if there is one).

I began this work with new context in the spec - that a user has
classified and therefore will have a UserProjectPreference (which
should include the project_id of the most recent classification). The
spec then says that it expects the last project to be included in the
email (needs to be updated to include project name).

Wrote some mailer code so this test passes. I could not get the
project_id or project name from the UserProjectPreference however and
upon inspection of 'last_project' in the spec, it is returning an empty
object and I'm not sure why.

For @camallen to look at