Alter text in dormant mailer if upp is true
If a user has classified before, then we want the dormant mailer to have
the option of linking them to their last project.

- If the user does have a previous classification then user project
preference (upp) will the true. If so then the id of the last project
classified is passed to @last_project in dormant user mailer.

- In the mailer text, if @last_project holds a project id then the
display name for that project and some surrounding text is given.
Otherwise it is left out and only a link to the main project page is
given.

- I will also add the project slug for the @last_project but not in this
commit.