Alter the output of dormant mailer
- The text for the dormant mailer emails is different depending on
whether a user has classified before (has a user project preference).
The addition of more changes between the two email types has meant it
made more sense to split the emails up completely in the mailer, despite
some of the text overlapping. This is much clearer to read than before.

- Also changed the default mailer layout to false and put in specific
unsubscribe notices depending on which email it is

- Checked the preview and looks good