Remove safe navigator and instance variable
Addressing a couple of comments made about a previous commit on this
file:

- removed the safe navigator to check if the user exists and recieves
emails. Uneeded to check if user exists as if the user does not exist
then an exception will be raised. This should never happen so if it does
it would be good if the exception is raised instead of being passed
over.

- Conversion of the instance variable @user to a normal variable