[WIP] begin looking for inactive users (#2882)
* begin looking for inactive users

* remove based on linter warning

* find all dormant users

use find_each to efficiently page through the user table checking the state of their last valid sign in time.

* add specs for explicit signin gaps

* fix inconsistent indentation

* Added classification check for dormant user

Updated the definition of a dormant user as someone who had not logged in for 5 days AND has not made any classifications in the last 5 days

* Added mailer file + spec for emailing dormant users

Want to send follow up emails to inactive users to encourage them to come back to the zooniverse

* added text for dormant users email

* added a worker file for scheduling emails to dormant users

* Updated mailer and spec for emailing dormant users

* added a preview file for the dormant user mailer

* changed name of the dormant user mailer method to email_dormant_user

* Begun worker to send mail to dormant users

So far the worker only sends an email to a user

* Freeze the subject string for the dormant mailer

Convert the default subject string into a frozn string

* Remove unnecessary comments

* Correct indentation error

* Remove unnecessary string conversion

* Improve syntax using the safe navigation operator

* move to checking latest lifecycled upp

non-classifying users have not lifecyled user projec prefs, this is a 
much faster query than classifications or recents. Long term this should 
be moved to an attribute on the user that tracks their last 
classification id instead of this proxy upp model.

* remove uneeded require

this was causing errors in the lifecycling spec

* avoid operation constant redefinition

allow the default operation class api_user to be missing as not all 
operations need an api_user. Any operations that uses the api_user will 
raise errors if the calling context doesn't supply correct param.

* update comment on the use of the UPP

* Add mailer preview paths to development config file

By default rails looks for mailer preview files in
test/mailers/previews/. However we have our previews in
spec/mailers/previews so I have updated the config file for the
development environment to point to the location of the previews in
spec.

* Make dormant user mailer preview work

- Updated the localhost URL that the preview can be viewed at during
development

-  Implemented the the method that calls DormantUserMailer

* Begin work to add last project to dormant mailer

When sending emails to dormant users it would be good to include a link
to the last project they've classified on (if there is one).

I began this work with new context in the spec - that a user has
classified and therefore will have a UserProjectPreference (which
should include the project_id of the most recent classification). The
spec then says that it expects the last project to be included in the
email (needs to be updated to include project name).

Wrote some mailer code so this test passes. I could not get the
project_id or project name from the UserProjectPreference however and
upon inspection of 'last_project' in the spec, it is returning an empty
object and I'm not sure why.

For @camallen to look at

* Update dormant mailer text

Updated the text that will be sent out to dormant users following
discussions with Grant. I have also updated the unsubscribe text to
match the default message.

* Remove unnecessary hash in unsubscribe link§§

* Use the default mailer layout

Changed the default layout to true for this mailer and so in this commit
I have removed the unsubscribe information from dormant mailer as this
will be pulled in from the default layout.

* Create user project preference for context in spec

For the context that a user has previously classified, create a user
project preference. Under this condition, it is expected that the last
user project name will be in the mailer.

* Alter text in dormant mailer if upp is true

If a user has classified before, then we want the dormant mailer to have
the option of linking them to their last project.

- If the user does have a previous classification then user project
preference (upp) will the true. If so then the id of the last project
classified is passed to @last_project in dormant user mailer.

- In the mailer text, if @last_project holds a project id then the
display name for that project and some surrounding text is given.
Otherwise it is left out and only a link to the main project page is
given.

- I will also add the project slug for the @last_project but not in this
commit.

* Email dormant users their most recent project

The dormant user mailer now includes a link to the users most recent
project they have contributed to. If the user has not made any
classifications then just a link to the main project page is given.

Also updated the mailer body and subject text in this commit.

* Alter routing in development mode

* Make factory bot gem available in development mode

I want to use the factory bot gem in dev mode when previewing mailers.
Previously this gem was only being loaded in testing mode.

* Re-order Gemfiles alphabetically

* Correct indentation

* Use the safe operator

* Increase dormant user window to 14 days

This is the time period after which if a user has not made a
classification then they will be marked as a dormant user.

* Remove blank line

* Alter the output of dormant mailer

- The text for the dormant mailer emails is different depending on
whether a user has classified before (has a user project preference).
The addition of more changes between the two email types has meant it
made more sense to split the emails up completely in the mailer, despite
some of the text overlapping. This is much clearer to read than before.

- Also changed the default mailer layout to false and put in specific
unsubscribe notices depending on which email it is

- Checked the preview and looks good

* Add further conditions for dormant user

Ensure that dormant users:

- Ignore disabled users
- Ensure dormant user has a valid email

* Add subset selection method for user

Subset users based on their id number, in this case if the user id ends
in 5.

* Add worker to that gets all dormant users

This worker get all dormant users within a subset and passes each one to
the dormant user mailer worker

* Remove blank line

* Enforce normal case for variable numbers

* Correct indentation

* Correct typo in variable name

* Add comment about why default route can't be there in dev mode

* Exclude users who have opted out of general emails

* Remove safe navigator and instance variable

Addressing a couple of comments made about a previous commit on this
file:

- removed the safe navigator to check if the user exists and recieves
emails. Uneeded to check if user exists as if the user does not exist
then an exception will be raised. This should never happen so if it does
it would be good if the exception is raised instead of being passed
over.

- Conversion of the instance variable @user to a normal variable

* Remove comment

* Add instant unsubscribe link

* Remove ternary operator for emails

* Improve readability

* Raise error if get dormant users is run on staging

* Fix rebase

* add space for short circut

* remove useless attr_reader setups

* switch the erro class def to it's own line

* add clearer where clause name

* test custom user unsubscribe link