Update the urls to contain google analytic codes
utm_source: 'dormant'

utm_campaign: depends is the email is for a user with a last user
project preference, so 'classification' or no last upp so 'no_classification'

utm_content: depends on which url was followed. 'a' is the all projects
page, 'b' is a specific project link, 'c' is the talk page link.