Ensure dormant mailers are not directing people to inactive or complete projects (#2952)
* Add first test for non-launch approved projects

If a users last project preference is not launch approved then we want
to send them a general email and not a project-specific email.

* Remove unnecessary parentheses

* Spec sets up a non-launch approved project

Corrected the spec test so that it sets up the user project preference
(i.e. the last project) as launch_approved = false. In this context, we
don't want to send users a last project specific email

* Check if the last project is launch approved

* Add a test for project completeness

* Send a general email if the project is complete

If the dormant user's last project (project preference) is completed
(has a completness of 1.0) then send the non-specific email, i.e don't
direct them back to the completed project.

* Include a thank you message if the last project is complete

* Update tests and finish the bulk of the email

* Add google analytic codes

Assign specific google analytic codes for users who have classified
before but whose last classified project is complete:
@dormant_with_last_project_complete_ga_code

* Extract duplicate logical combination to variable

* Use safe navigator

* Remove link to the general talk page

Directing someone to the general talk page is not very helpful for
getting them back on the zooniverse. Only include the link in the
project-specific emails.

* Refactor tests