allow user intervention column name change
intervention_notifications changes to interventions in 
https://github.com/zooniverse/Panoptes/pull/3010

This is a pre-deploy commit to allow the atttributes to serialize with 
the correct value until the column name, code and user serializer schema 
is changed. The plan is to deploy this to the API, migrate the db, 
deploy the new code and serializer schema and update the front end to 
use the new attributes. Then remove this vestigal code in another future 
PR.