Fix code load order issue with specs
In certain spec runs, and when you only run the AwsAdapter spec in
isolation, autoloading wouldn't load the AbstractAdapter before getting
to a line in the spec which references this error class. Because Rails
has no way of knowing it's defined in the AbstractAdapter file rather
than the file defining the base MediaStorage module, you end up with an
undefined constant error.