Build missing workflow version on the fly
If the most recent version is requested, but for some reason does not
appear in the WorkflowVersions table, we can instatiate this version on
the fly.

This happens because when we backfilled the workflow versions, we only
backfilled any versions that were referenced by classifications. Some
workflows had no classifications at that time, and so they never got any
workflow versions backfilled to them. While we should go and backfill
those anyway, at this point this is somewhat tricky to do.