Export catch nomethoderror (#3163)
* Catch NoMethodError when generating export

`workflow_at_version` can be nil. I assume there is an underlying issue
that causes that, but this does not attempt to fix it.

Regardless of the underlying problem, the export should not silently
fail.

* Build missing workflow version on the fly

If the most recent version is requested, but for some reason does not
appear in the WorkflowVersions table, we can instatiate this version on
the fly.

This happens because when we backfilled the workflow versions, we only
backfilled any versions that were referenced by classifications. Some
workflows had no classifications at that time, and so they never got any
workflow versions backfilled to them. While we should go and backfill
those anyway, at this point this is somewhat tricky to do.

* Add rake task to fill in workflow versions if none were created at all

* Move error handling up and more specific