remove extra priority filter for update query
We can't import duplicate SMS records as the DB has a uniq index for subject_id + subject_set_id.

Thus if the import runs then the SMS record is new and will have a NULL priority value by default.

thus we can rely on the SMS.id field for selectivity on the update query