check the sms resources exist attempting an update
sms resources are created in a transaction via `add_relation` and then the worker is fired. DB isolation levels mean that this worker will most likely run before the transaction is commited, thus the sms record won't be visible to the update statement.

this PR will ensure we raise the error and use sidekiqs retry attempts to re-reun the update after transaction commit.