update structure.sql
for some reason pg_dump now adds versions of db and dump tool
and removes the tablespace

I can't find any reason why this is happening but i've tested this on various installs (docker, brew, manual install) so i assume it's pg_dump related.

Rails code hasn't changed for a while and it wraps pg_dump.

this should be a one of change, pity it breaks the diff though