Add subject_group data model (#3511)
* add subject group and join table migrations

* update structure.sql

for some reason pg_dump now adds versions of db and dump tool
and removes the tablespace

I can't find any reason why this is happening but i've tested this on various installs (docker, brew, manual install) so i assume it's pg_dump related.

Rails code hasn't changed for a while and it wraps pg_dump.

this should be a one of change, pity it breaks the diff though

* add subject group model with habtm subject

* remove wip

* fix linter errors

* switch to has_many through model

add a resource of SubjectGroupMember to record the link of a subject to a subject group and the display order of this resource

* add SubjectGroupMember resource

* fixing specs on subject_group model

* use true to force the association to load

* re-run migration for subject group resource creation

* add context json field or subject group

record information about when / how the subject group was created, e.g. subject selection information

* add key to subject_group

track the known subject-ids that comprise this subject group.

This key field will be used to uniquely identify the subject group and the associated subject for talk etc.

* ensure subject group key is in order

uniquely identify this subject group based on the selection order of the subject group members

different groups will have different key signatures and orders of showing the subjects

* add unique index to subject_group key

* set key before save, remove validation checks

rely on the db to enforce this key presence / uniqueness

* add group_subject reference to grouped_subject model

* add group_subject association

* simplify model validations

use service object to set the key and build the group_subject before save

* add subject group create operation

adds a service object to correctly build the subject group and associated records.

* remove subjects.subject_group has_many associations

avoid tracking the group via an association, track the group via the metadata object, this simplifies the export tooling as well.

* track the subject_group info in the group_subject metadata

allow the subject to have the information about the group for use in exports / downstream analysis, i.e. retirement etc