split default and www-panoptes ingress certificates (#3555)
ensure the www-panoptes ingress has a valid certificate. Putting all the dns domain on one cert clobbers if they aren't used on the same ingress, i.e. we lose www-panoptes.zooniverse.org on the cert.

I think this is related to the shim and the ingress `cert-manager.io` annotation but i'm not sure.

TBH It's explicit this way and most likely easier to track the certs and ingress domains as well.