avoid class level var overrides at spec runtime
the value of  SUBJECT_GROUP_MIN_SIZE env is used when the class is loaded / defined not at run time. Thus it depends on when the class is loaded as to what env value is set.

Avoid this env race condition setup and instead just test with more realistic factory data