improve project paused state transitions
move a project to a paused state if there is no subject sets on any of the active workflows (i.e. it's out of data).

Some project teams unlink subject sets when the workflow / project is finished but leave the workflow active, this will avoid these projects coming back to active.