remove sidetiq and sidetiq configs. remove redis.exists_return_integer flag (sidekiq-cron and sidekiq versions has this taken care of)
Redis.exists_return_integer = false was added because sidetiq expects redis.exists to return a Boolean instead of an integer. redis gem 4.5.1 changed the return value to integer but also added the option to return boolean as long as we set this flag somewhere.

Since we are moving away from sidetiq, we do not need this flag anymore because scheduler (sidekiq-cron) and sidekiq versions have taken care of this redis update within their gem