switch to proportional project completeness metric
taking the average of all workflow completion metrics doesn't factor the proportion of the subjects that each workflow is responsible for.

switching to a proportional metric that factors in the number of subjects each workflow has ensure we have a more accurate completeness metric that takes these subject proportions into account as well as edge cases like one training workflow have 0 subjects but a 0 completess metric where an average doesn't work.